Remove periodic images from the system we pass to metatomic
Many ML models are very slow when running on all atoms in the system,
and because we remap the pairs, periodic images end up as isolated atoms
for which we throw away the prediction.

Doing some more processing on the LAMMPS side only adds minimal overhead
while removing many potential atomic centers and making the ML models
a lot faster.